requirements.txt
# Core dependencies
python>=3.10
pyautogui>=0.9.54
pywinauto>=0.6.8
opencv-python>=4.8.0
pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0

# LLM Clients
openai>=1.0.0
anthropic>=0.8.0
qwen-llm>=0.1.0

# Web/API
requests>=2.31.0
httpx>=0.25.0
websockets>=12.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-image>=0.22.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.5.0

# Logging
loguru>=0.7.0

# Type hints
typing-extensions>=4.8.0